+++ src/main/java/pl/schoolmanagementsystem/common/mark/dto/MarkWithTwoFields.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class MarkWithTwoFields {
 
-    private int mark;
+    private byte mark;
     private String subject;
 
 }
